requirements.txt
flask==2.0.1
flask-cors==3.0.10
requests==2.26.0
python-dotenv==0.19.0
gunicorn==20.1.0
certifi==2021.10.8
urllib3==1.26.7